+++ services/imagesearch/src/main/java/com/huaweicloud/sdk/imagesearch/v2/model/SearchInfo.java
@@ -3,7 +3,6 @@ package com.huaweicloud.sdk.imagesearch.v2.model;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-import java.math.BigDecimal;
 import java.util.Objects;
 import java.util.function.Consumer;
 
@@ -15,24 +14,24 @@ public class SearchInfo {
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "total_num")
 
-    private BigDecimal totalNum;
+    private Integer totalNum;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "return_num")
 
-    private BigDecimal returnNum;
+    private Integer returnNum;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "search_time")
 
-    private BigDecimal searchTime;
+    private Integer searchTime;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "last_item")
 
     private SearchAfterParam lastItem;
 
-    public SearchInfo withTotalNum(BigDecimal totalNum) {
+    public SearchInfo withTotalNum(Integer totalNum) {
         this.totalNum = totalNum;
         return this;
     }
@@ -41,15 +40,15 @@ public class SearchInfo {
      * 搜索结果总数。
      * @return totalNum
      */
-    public BigDecimal getTotalNum() {
+    public Integer getTotalNum() {
         return totalNum;
     }
 
-    public void setTotalNum(BigDecimal totalNum) {
+    public void setTotalNum(Integer totalNum) {
         this.totalNum = totalNum;
     }
 
-    public SearchInfo withReturnNum(BigDecimal returnNum) {
+    public SearchInfo withReturnNum(Integer returnNum) {
         this.returnNum = returnNum;
         return this;
     }
@@ -58,15 +57,15 @@ public class SearchInfo {
      * 返回结果总数。
      * @return returnNum
      */
-    public BigDecimal getReturnNum() {
+    public Integer getReturnNum() {
         return returnNum;
     }
 
-    public void setReturnNum(BigDecimal returnNum) {
+    public void setReturnNum(Integer returnNum) {
         this.returnNum = returnNum;
     }
 
-    public SearchInfo withSearchTime(BigDecimal searchTime) {
+    public SearchInfo withSearchTime(Integer searchTime) {
         this.searchTime = searchTime;
         return this;
     }
@@ -75,11 +74,11 @@ public class SearchInfo {
      * 搜索过程耗时，单位为毫秒。
      * @return searchTime
      */
-    public BigDecimal getSearchTime() {
+    public Integer getSearchTime() {
         return searchTime;
     }
 
-    public void setSearchTime(BigDecimal searchTime) {
+    public void setSearchTime(Integer searchTime) {
         this.searchTime = searchTime;
     }
